+++ CHANGELOG.md
@@ -2,7 +2,14 @@
 
 ## 0.2.0 (??.??.202?)
 
-*Nothing yet*
+### Added:
+
+ + method `get_me` to class `Site`;
+ + method `get_my_unread_inbox` to class `Site`;
+
+### Changed:
+
+ + error messages, it was: `"{message}" on URL {url} (status code {status_code})`, it became: `{message} (status code {status_code})`
 
 ## 0.1.1 (25.09.2023)